requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.2.2
Django==4.2.11
djangorestframework==3.15.1
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
sib-api-v3-sdk==7.6.0
six==1.16.0
sqlparse==0.5.0
typing-extensions==4.11.0
urllib3==2.2.1
